backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11+
# To update, run: pip-compile requirements.in
#
# SECURITY: These versions are pinned for reproducible builds
# Run pip-compile --upgrade periodically to update
#

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
# starlette version managed by fastapi

# Data validation
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic-core==2.27.2

# Environment
python-dotenv==1.0.1
python-multipart==0.0.20

# Security
PyJWT==2.10.1
bcrypt==4.2.1
cryptography==44.0.0

# Database - SQLAlchemy with async support
sqlalchemy==2.0.37
alembic==1.14.1
aiosqlite==0.20.0
asyncpg==0.30.0
greenlet==3.1.1

# HTTP client
httpx==0.28.1
httpcore==1.0.7
h11==0.14.0
anyio==4.8.0
sniffio==1.3.1
certifi==2024.12.14
idna==3.10

# AI/LLM
openai==1.59.9
distro==1.9.0
jiter==0.8.2
tqdm==4.67.1

# Cache
redis==5.2.1

# Rate limiting
slowapi==0.1.9
limits==3.14.1

# Utilities
click==8.1.8
Mako==1.3.8
MarkupSafe==3.0.2
typing_extensions==4.12.2
annotated-types==0.7.0
email-validator==2.2.0
dnspython==2.7.0

# ASGI/Uvicorn extras
uvloop==0.21.0; sys_platform != "win32"
httptools==0.6.4
watchfiles==1.0.4
websockets==14.2
PyYAML==6.0.2
